refractor de communication!
