+++ Communication/receiver/vision_receiver.py
@@ -4,12 +4,13 @@
     serveur de vision.
 """
 
-from .udp_server import PBPacketReceiver
-from .protobuf import messages_robocup_ssl_wrapper_pb2 as ssl_wrapper
+from RULEngine.Communication.protobuf import messages_robocup_ssl_wrapper_pb2 as ssl_wrapper
+from RULEngine.Communication.util.protobuf_packet_receiver import ProtobufPacketReceiver
 
-class Vision(PBPacketReceiver):
+
+class VisionReceiver(ProtobufPacketReceiver):
     """ Initie le serveur de vision, semble superflue. """
     # FIXME: est-ce réellement utile? la classe ne semble aucunement générique
 
     def __init__(self, host="224.5.23.2", port=10020):
-        super(Vision, self).__init__(host, port, ssl_wrapper.SSL_WrapperPacket)
+        super(VisionReceiver, self).__init__(host, port, ssl_wrapper.SSL_WrapperPacket)
