+++ Communication/util/serial_protocol.py
@@ -8,6 +8,7 @@ ESCAPEBYTE = b'\x7D'
 SPEEDCOMMAND_ID = 1
 PIDCOMMAND_ID = 2
 
+
 def create_speed_command(x, y, theta, id):
 
     packet = struct.pack('<BBfff', id, SPEEDCOMMAND_ID, x, y, theta)
