+++ Communication/util/threaded_udp_server.py
@@ -1,11 +1,11 @@
-#Under MIT License, see LICENSE.txt
-#!/usr/bin/python
+# Under MIT License, see LICENSE.txt
 
 from socketserver import ThreadingMixIn, UDPServer
 import threading
 import socket
 import struct
 from ipaddress import ip_address
+
 
 class ThreadedUDPServer(ThreadingMixIn, UDPServer):
 
@@ -28,11 +28,3 @@ class ThreadedUDPServer(ThreadingMixIn, UDPServer):
         server_thread = threading.Thread(target=self.serve_forever)
         server_thread.daemon = True
         server_thread.start()
-
-
-def udp_socket(host, port):
-    skt = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    skt.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-    connection_info = (host, port)
-    skt.connect(connection_info)
-    return skt
