+++ Communication/util/udp_socket.py
@@ -1,0 +1,11 @@
+# Under MIT License, see LICENSE.txt
+
+import socket
+
+
+def udp_socket(host, port):
+    skt = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    skt.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+    connection_info = (host, port)
+    skt.connect(connection_info)
+    return skt
